capture: give the deep read's principal its own file
deepread.go crossed the 500-line cap taking on the claim-derived principal. Who
a read acts as is its own subject — the worker runs on a job, but everything it
writes belongs to the human the dossier row names, and that distinction is what
the new file is about.
