+++ backend/internal/compose/deepread.go
@@ -25,7 +25,6 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"strings"
 	"time"
 
 	"github.com/jackc/pgx/v5/pgxpool"
@@ -37,7 +36,6 @@ import (
 	"github.com/gradionhq/margince/backend/internal/platform/webread"
 	"github.com/gradionhq/margince/backend/internal/shared/apperrors"
 	"github.com/gradionhq/margince/backend/internal/shared/kernel/ids"
-	"github.com/gradionhq/margince/backend/internal/shared/kernel/principal"
 )
 
 // SiteDeepReadArgs is one queued deep read. The args carry everything the
@@ -143,43 +141,6 @@ func (w *siteDeepReadWorker) Timeout(*river.Job[SiteDeepReadArgs]) time.Duration
 // both its configured crawl budget and the time reserved to close the dossier.
 func (w *siteDeepReadWorker) reclaimAfter() time.Duration {
 	return w.Timeout(nil) + time.Minute
-}
-
-// run is the whole deep read, River-agnostic so tests drive it directly.
-// Retry semantics rest on BeginSiteRead's CAS: any terminal outcome
-// (done, partial, failed) leaves the dossier past "queued", so a River
-// retry — including one after a recorded failure — CAS-misses and
-// no-ops. One honest outcome per dossier, no zombie re-crawls; reading
-// the site again is a human's next start, never an automatic retry.
-// deepReadWorkerCtx attaches the worker's principal, workspace and correlation
-// onto the job context — the values every store write (and terminalCtx) needs.
-//
-// The requester it carries comes from the job payload, because the claim that
-// yields the authoritative one needs a context to be made at all. It is
-// replaced by withClaimedRequester the moment the row is in hand, so nothing
-// the worker writes is attributed on the payload's word.
-func deepReadWorkerCtx(ctx context.Context, args SiteDeepReadArgs) context.Context {
-	return withClaimedRequester(
-		principal.WithWorkspaceID(ctx, args.WorkspaceID), args.RequestedBy, args.SiteReadID)
-}
-
-// withClaimedRequester stamps the principal every store write is attributed to.
-// The human named here owns what this read creates, so it must be the one the
-// DOSSIER ROW records: a payload that disagreed would hang another person's
-// name on the rows, which no later gate would catch — provenance is written
-// once and never re-derived.
-func withClaimedRequester(ctx context.Context, requestedBy string, readID ids.UUID) context.Context {
-	requester := requestedByUserID(requestedBy)
-	ctx = principal.WithActor(ctx, principal.Principal{
-		Type:       principal.PrincipalSystem,
-		ID:         "agent:deepread",
-		UserID:     requester,
-		OnBehalfOf: requester,
-	})
-	if readID.IsZero() {
-		return principal.WithCorrelationID(ctx, ids.NewV7())
-	}
-	return principal.WithCorrelationID(ctx, readID)
 }
 
 func (w *siteDeepReadWorker) run(ctx context.Context, args SiteDeepReadArgs) error {
@@ -499,20 +460,3 @@ func (w *siteDeepReadWorker) finish(ctx context.Context, readID ids.UUID, status
 // (deepreadaccept.go). Distinct from the quick scrape's "enrich": a deep
 // read's acceptance also lands category facts.
 const deepReadProposalKind = "deepread"
-
-// requestedByUserID recovers the human uuid behind a "human:<uuid>"
-// requested_by so the staged proposal carries OnBehalfOf. A requester
-// without a recoverable uuid yields the zero uuid — the approval's
-// on_behalf_of is then honestly NULL rather than the read failing over
-// provenance.
-func requestedByUserID(requestedBy string) ids.UUID {
-	_, raw, found := strings.Cut(requestedBy, ":")
-	if !found {
-		return ids.UUID{}
-	}
-	id, err := ids.Parse(raw)
-	if err != nil {
-		return ids.UUID{}
-	}
-	return id
-}
